==> 3-RunBacktest copy.ipynb <==
## Import Required Libraries

We'll import all necessary Python libraries for financial data analysis, portfolio optimization, and visualization.

import warnings
import os
#from datetime import datetime, timedelta
import json
import pandas as pd
#import numpy as np
#import random
#import sys
from IPython.display import display, Markdown
#sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'AssetPricingPortfolio')))

from Backtester.helpers import read_datasets_data,recursive_serialize

from Backtester.BacktestFramework import BacktraderPortfolioBacktest

from Backtester.Strategy_Core import EqualWeightStrategy #, InverseVolatilityStrategy,QuintileStrategy,GMVPStrategy,MarkowitzStrategy,VanillaRiskParityStrategy,MaximumSharpeRatioStrategy,MostDiversifiedStrategy,MaximumDecorrelationStrategy
from Backtester.Strategy_AAA import AdaptiveAssetAllocationStrategy
from Backtester.Strategy_dual_momentum import GlobalEquitiesMomentumStrategy

warnings.filterwarnings('ignore')

# Define settings, load universe data and prepare resampled datasets for backtest

# define backtest settings
universe_name = "selection3" 
test_name = "GEM_LP"

#load the test data

universe_folder_path = os.path.join("data",universe_name)
datasets_folder_path = os.path.join("data",universe_name,test_name,"datasets")
test_folder_path = os.path.join("data",universe_name,test_name)

with open(os.path.join(universe_folder_path, 'universe_settings.json'), 'r') as f:
    universe_settings = json.load(f)

with open(os.path.join(test_folder_path, 'datasets_settings.json'), 'r') as f:
    datasets_settings = json.load(f)

with open(os.path.join(test_folder_path, 'test_settings.json'), 'r') as f:
    test_settings = json.load(f)

universe_info_df = pd.read_csv(os.path.join(universe_folder_path, "universe_info.csv"), header=[0], index_col=0, parse_dates=True)
universe_data = pd.read_csv(os.path.join(universe_folder_path, "universe_data.csv"), header=[0,1], index_col=0, parse_dates=True)

datasets_info, my_dataset_list = read_datasets_data(test_folder_path)

backtest_settings = {
    'lookback':datasets_settings['lookback_periods'],
}

print(backtest_settings)  


backtest_settings['lookback'] = 21*6+1  # in days
backtest_settings['rebalance_every'] = 7  # in days
backtest_settings['initial_cash'] = 100000
backtest_settings['commission'] = 0.001  # 0.1% commission
backtest_settings['short_interest'] = 0.03  # 3% annual short interest
backtest_settings['interest_long'] = False  # Only charge shorts
backtest_settings['checksubmit'] = False

display(Markdown(f"### Backtest Settings\n```json\n{json.dumps(backtest_settings, indent=4)}\n```"))

try:
    with open(os.path.join(test_folder_path, 'strategies.json'), 'r') as f:
        existing_strategies = json.load(f)
    #display(Markdown(f"### Existing Strategies\n```json\n{json.dumps(existing_strategies, indent=4)}\n```"))
    display(existing_strategies)

except FileNotFoundError:
    existing_strategies = []

bt_strategies = {
    #"EqualWeight": (EqualWeightStrategy, {
    #    'buy_slippage_buffer': 0.01,
    #}),
    #'GEM_M1': (GlobalEquitiesMomentumStrategy, {
    #    'buy_slippage_buffer': 0.03,
    #    'use_stops': False,
    #    'trailing_stop': True,
    #    'trail_percent': 0.05,  # 5% trailing stop
    #    'verbose': False,
    #    'portfolio_func_kwargs': {
    #        'allow_short': False,
    #        'momentum_periods': [10, 21, 21*3, 21*6],  # Days, not months!
    #        'min_positive_periods': 3,
    #        'treasury_threshold': 0.0,
    #        'maximum_positions': 6,
    #        'momentum_method': 'simple',
    #        'bb_period': 20, # used if momentum_method is 'bollinger'
    #        'bb_std': 2.0, # used if momentum_method is 'bollinger'
    #        'risk_free_asset': 'cash',  # 'cash' (default) or symbol like 'BIL', 'SHY'
    #    }
    #}),
    #'GEM2': (GlobalEquitiesMomentumStrategy, {
    #    'buy_slippage_buffer': 0.03,
    #    'use_stops': True,
    #    'trailing_stop': True,
    #    'trail_percent': 0.05,  # 5% trailing stop
    #    'verbose': False,
    #    'portfolio_func_kwargs': {
    #        'allow_short': False,
    #        'momentum_periods': [10, 21, 21*3, 21*6],  # Days, not months!
    #        'min_positive_periods': 3,
    #        'treasury_threshold': 0.0,
    #        'maximum_positions': 6,
    #        'momentum_method': 'bollinger',
    #        'bb_period': 20, # used if momentum_method is 'bollinger'
    #        'bb_std': 2.0, # used if momentum_method is 'bollinger'
    #        'risk_free_asset': 'cash',  # 'cash' (default) or symbol like 'BIL', 'SHY'
    #    }
    #}),
    #'GEM3': (GlobalEquitiesMomentumStrategy, {
    #    'buy_slippage_buffer': 0.03,
    #    'use_stops': True,
    #    'trailing_stop': True,
    #    'trail_percent': 0.05,  # 5% trailing stop
    #    'verbose': False,
    #    'portfolio_func_kwargs': {
    #        'allow_short': False,
    #        'momentum_periods': [10, 21, 21*3, 21*6],  # Days, not months!
    #        'min_positive_periods': 3,
    #        'treasury_threshold': 0.0,
    #        'maximum_positions': 6,
    #        'momentum_method': 'simple',
    #        'bb_period': 20, # used if momentum_method is 'bollinger'
    #        'bb_std': 2.0, # used if momentum_method is 'bollinger'
    #        'risk_free_asset': 'cash',  # 'cash' (default) or symbol like 'BIL', 'SHY'
    #    }
    #}),
    'GEM4': (GlobalEquitiesMomentumStrategy, {
        'buy_slippage_buffer': 0.03,
        'use_stops': False,
        'trailing_stop': True,
        'trail_percent': 0.05,  # 5% trailing stop
        'verbose': False,
        'portfolio_func_kwargs': {
            'allow_short': False,
            'momentum_periods': [10, 21, 21*3, 21*6],  # Days, not months!
            'min_positive_periods': 3,
            'treasury_threshold': 0.0,
            'maximum_positions': 6,
            'momentum_method': 'bollinger',
            'bb_period': 20, # used if momentum_method is 'bollinger'
            'bb_std': 2.0, # used if momentum_method is 'bollinger'
            'risk_free_asset': 'cash',  # 'cash' (default) or symbol like 'BIL', 'SHY'
        }
    }),
    #'GEM5_noComission': (GlobalEquitiesMomentumStrategy, {
    #    'buy_slippage_buffer': 0.03,
    #    'use_stops': True,
    #    'trailing_stop': True,
    #    'trail_percent': 0.05,  # 5% trailing stop
    #    'verbose': False,
    #    'portfolio_func_kwargs': {
    #        'allow_short': False,
    #        'momentum_periods': [10, 21, 21*3, 21*6],  # Days, not months!
    #        'min_positive_periods': 4,
    #        'treasury_threshold': 0.0,
    #        'maximum_positions': 10,
    #        'momentum_method': 'simple',
    #        'bb_period': 20, # used if momentum_method is 'bollinger'
    #        'bb_std': 2.0, # used if momentum_method is 'bollinger'
    #        'risk_free_asset': 'cash',  # 'cash' (default) or symbol like 'BIL', 'SHY'
    #    }
    #}),
    #"AAA1": (AdaptiveAssetAllocationStrategy, {
    #    'buy_slippage_buffer': 0.03,
    #    'use_stops': False,
    #    'trailing_stop': False,
    #    'trail_percent': 0.05,
    #    'verbose': False,
    #    'portfolio_func_kwargs': {
    #        'momentum_method': 'simple',
    #        'momentum_top_n': 10, 
    #        'momentum_lookback': 21*6,
    #        'covariance_lookback': 21,
    #    }
    #}),
    #"AAA_M1": (AdaptiveAssetAllocationStrategy, {
    #    'buy_slippage_buffer': 0.03,
    #    'use_stops': False,
    #    'trailing_stop': False,
    #    'trail_percent': 0.05,
    #    'verbose': False,
    #    'portfolio_func_kwargs': {
    #        'momentum_method': 'simple',
    #        'momentum_top_n': 10, 
    #        'momentum_lookback': 21*6,
    #        'covariance_lookback': 21,
    #    }
    #}),
    #"AAA2_noComission": (AdaptiveAssetAllocationStrategy, {
    #    'buy_slippage_buffer': 0.03,
    #    'use_stops': True,
    #    'trailing_stop': True,
    #    'trail_percent': 0.05,  
    #    'verbose': False,
    #    'portfolio_func_kwargs': {
    #        'momentum_method': 'simple',
    #        'momentum_top_n': 10,
    #        'momentum_lookback': 21*6,
    #        'covariance_lookback': 21,
    #    }
    #}),
    #"AAA3": (AdaptiveAssetAllocationStrategy, {
    #    'buy_slippage_buffer': 0.03,
    #    'use_stops': True,
    #    'trailing_stop': True,
    #    'trail_percent': 0.05,  
    #    'verbose': False,
    #    'portfolio_func_kwargs': {
    #        'momentum_method': 'bollinger',
    #        'bb_period': 20,
    #        'bb_std': 2.0,
    #        'momentum_top_n': 10,
    #        'momentum_lookback': 21*6,
    #        'covariance_lookback': 21,
    #    }
    #}),
    #"AAA4": (AdaptiveAssetAllocationStrategy, {
    #    'buy_slippage_buffer': 0.03,
    #    'use_stops': False,
    #    'trailing_stop': False,
    #    'trail_percent': 0.05,  
    #    'verbose': False,
    #    'portfolio_func_kwargs': {
    #        'momentum_method': 'bollinger',
    #        'bb_period': 20,
    #        'bb_std': 2.0,
    #        'momentum_top_n': 10,
    #        'momentum_lookback': 21*6,
    #        'covariance_lookback': 21,
    #    }
    #}),
}

#existing_strategies.append(recursive_serialize(bt_strategies))
#with open(os.path.join(test_folder_path, 'strategies.json'), 'w') as f:
##    json.dump(existing_strategies, f, indent=4)

bt_backtest = BacktraderPortfolioBacktest(
    strategies=bt_strategies, 
    datasets=my_dataset_list,
    rebalance_every=backtest_settings['rebalance_every'],  # in days
    lookback=int(backtest_settings['lookback']),  # in days
    initial_cash=backtest_settings['initial_cash'],
    commission=backtest_settings['commission'],
    short_interest=backtest_settings['short_interest'],
    interest_long=backtest_settings['interest_long'],
    #checksubmit=False,
)

# Run the backtest
backtest_results = bt_backtest.run_backtest()

print("\nBacktrader backtesting with warmup period completed successfully!")
print(f"Lookback period: {bt_backtest.lookback} days")
print(f"Rebalancing frequency: {bt_backtest.rebalance_every} days")

# save results

serializable_results = recursive_serialize(backtest_results)

results_folder_path = os.path.join(test_folder_path, 'results')
os.makedirs(results_folder_path, exist_ok=True)

for strategy, datasets in serializable_results.items():
    strategy_results = {}
    strategy_results['Strategy'] = strategy
    strategy_results['Datasets'] = datasets
    with open(os.path.join(results_folder_path, f'{strategy}_results.json'), 'w') as f:
        json.dump(strategy_results, f, indent=2)

"""
    #"Quintile": (QuintileStrategy, {
    #    'buy_slippage_buffer': 0.07,
        #'checksubmit': False, 
    #}),
    "MVP": (MarkowitzStrategy, {
        'buy_slippage_buffer': 0.07,
        'portfolio_func_kwargs': {'allow_short': False}
        #'checksubmit': False,
    }),

    "GMVP": (GMVPStrategy, {
        'buy_slippage_buffer': 0.05,
        'portfolio_func_kwargs': {'allow_short': False}
    }),
    "IVP": (InverseVolatilityStrategy, {
        'buy_slippage_buffer': 0.04,
    }),
    "RPP": (VanillaRiskParityStrategy, {
        'buy_slippage_buffer': 0.04,
    }),
    "MSRP": (MaximumSharpeRatioStrategy, {
        'buy_slippage_buffer': 0.05,
        'portfolio_func_kwargs': {'allow_short': False}
    }),
    "MDP": (MostDiversifiedStrategy, {
        'buy_slippage_buffer': 0.04,
        'portfolio_func_kwargs': {'allow_short': False}
    }),
    "MDC": (MaximumDecorrelationStrategy, {
        'buy_slippage_buffer': 0.04,
        'portfolio_func_kwargs': {'allow_short': False}
    }),
    "EqualWeight": (EqualWeightStrategy, {
        'buy_slippage_buffer': 0.03,
    }),
"""
